[NAN-20] implement hubspot webhooks (#1407)
* [nan-20] parse new properties and set validation logic

• add tests to support new properties

* [nan-20] replaceMetada (set) and updateMetada

• update docs
• adds tests

* use connectionCreatedHook to make it easier to manage connection lifecycle

* [nan-20] progress on webhook and post connection

* [nan-20] webhook logic written in the sync script

* [nan-20] allow a webhook update to send a nango webhook once finished

* [nan-20] feedback updates

* [nan-20] fix spacing

* [nan-20] update test failure

* [nan-20] remove todo

* [nan-20] don't allow invalid webhooks through

* [nan-20] query for the connection using the identifier instead of looping over all connections

* [nan-20] observability

* [nan-20] add jira webhook handling

* [nan-20] pr feedback

* [nan-20] remove deprecated

* [nan-20] remove runner node_modules and dist

* [nan-20] update webhook logic

* [nan-20] add possibility for customer defined feature flag

* [nan-20] remove tsbuild file

* [nan-20] strip extra connection information

+++ packages/shared/lib/services/sync/run.service.ts
@@ -28,6 +28,8 @@ interface SyncRunConfig {
     integrationService: IntegrationServiceInterface;
     writeToDb: boolean;
     isAction?: boolean;
+    isInvokedImmediately?: boolean;
+    isWebhook?: boolean;
     nangoConnection: NangoConnection;
     syncName: string;
     syncType: SyncType;
@@ -51,6 +53,7 @@ export default class SyncRun {
     integrationService: IntegrationServiceInterface;
     writeToDb: boolean;
     isAction: boolean;
+    isInvokedImmediately: boolean;
     nangoConnection: NangoConnection;
     syncName: string;
     syncType: SyncType;
@@ -67,14 +70,17 @@ export default class SyncRun {
     stubbedMetadata?: Metadata | undefined = undefined;
 
     temporalContext?: Context;
+    isWebhook: boolean;
 
     constructor(config: SyncRunConfig) {
         this.integrationService = config.integrationService;
         this.writeToDb = config.writeToDb;
         this.isAction = config.isAction || false;
+        this.isWebhook = config.isWebhook || false;
         this.nangoConnection = config.nangoConnection;
         this.syncName = config.syncName;
         this.syncType = config.syncType;
+        this.isInvokedImmediately = Boolean(config.isAction || config.isWebhook);
 
         if (config.syncId) {
             this.syncId = config.syncId;
@@ -149,7 +155,7 @@ export default class SyncRun {
                 console.error(message);
             }
 
-            const errorType = this.isAction ? 'action_script_failure' : 'sync_script_failure';
+            const errorType = this.determineErrorType();
             return { success: false, error: new NangoError(errorType, message, 404), response: false };
         }
 
@@ -159,7 +165,7 @@ export default class SyncRun {
         if (!integrations[this.nangoConnection.provider_config_key] && !this.writeToDb) {
             const message = `The connection you provided which applies to integration "${this.nangoConnection.provider_config_key}" does not match any integration in the ${nangoConfigFile}`;
 
-            const errorType = this.isAction ? 'action_script_failure' : 'sync_script_failure';
+            const errorType = this.determineErrorType();
             return { success: false, error: new NangoError(errorType, message, 404), response: false };
         }
 
@@ -174,7 +180,7 @@ export default class SyncRun {
             if (!environment && !bypassEnvironment) {
                 const message = `No environment was found for ${this.nangoConnection.environment_id}. The sync cannot continue without a valid environment`;
                 await this.reportFailureForResults(message);
-                const errorType = this.isAction ? 'action_script_failure' : 'sync_script_failure';
+                const errorType = this.determineErrorType();
                 return { success: false, error: new NangoError(errorType, message, 404), response: false };
             }
 
@@ -223,7 +229,7 @@ export default class SyncRun {
                     const message = `Integration was attempted to run for ${this.syncName} but no integration file was found at ${integrationFilePath}.`;
                     await this.reportFailureForResults(message);
 
-                    const errorType = this.isAction ? 'action_script_failure' : 'sync_script_failure';
+                    const errorType = this.determineErrorType();
 
                     return { success: false, error: new NangoError(errorType, message, 404), response: false };
                 }
@@ -231,7 +237,7 @@ export default class SyncRun {
 
             let lastSyncDate: Date | null | undefined = null;
 
-            if (!this.isAction) {
+            if (!this.isInvokedImmediately) {
                 if (!this.writeToDb) {
                     lastSyncDate = optionalLastSyncDate;
                 } else {
@@ -310,7 +316,8 @@ export default class SyncRun {
                     syncData,
                     this.nangoConnection.environment_id,
                     this.writeToDb,
-                    this.isAction,
+                    this.isInvokedImmediately,
+                    this.isWebhook,
                     this.loadLocation,
                     this.input,
                     this.temporalContext
@@ -459,7 +466,7 @@ export default class SyncRun {
                                     const message = `There was a problem upserting the data for ${this.syncName} and the model ${model} with the error message: ${upsertResult?.error}`;
                                     await this.reportFailureForResults(message);
 
-                                    const errorType = this.isAction ? 'action_script_failure' : 'sync_script_failure';
+                                    const errorType = this.determineErrorType();
 
                                     return { success: false, error: new NangoError(errorType, message), response: result };
                                 }
@@ -479,7 +486,7 @@ export default class SyncRun {
                     } sync did not complete successfully and has the following error: ${errorMessage}`
                 );
 
-                const errorType = this.isAction ? 'action_script_failure' : 'sync_script_failure';
+                const errorType = this.determineErrorType();
 
                 return { success: false, error: new NangoError(errorType, errorMessage), response: result };
             }
@@ -491,12 +498,12 @@ export default class SyncRun {
     async finishSync(models: string[], syncStartDate: Date, version: string, totalRunTime: number, trackDeletes?: boolean): Promise<void> {
         let i = 0;
         for (const model of models) {
-            if (trackDeletes) {
+            if (!this.isWebhook && trackDeletes) {
                 await clearOldRecords(this.nangoConnection?.id as number, model);
             }
             const deletedKeys = trackDeletes ? await getDeletedKeys('_nango_sync_data_records', 'external_id', this.nangoConnection.id as number, model) : [];
 
-            if (trackDeletes) {
+            if (!this.isWebhook && trackDeletes) {
                 await syncUpdateAtForDeletedRecords(this.nangoConnection.id as number, model, 'external_id', deletedKeys);
             }
 
@@ -536,19 +543,21 @@ export default class SyncRun {
             // any changes while the sync is running
             // but if the sync date was set by the user in the integration script,
             // then don't override it
-            const override = false;
-            await setLastSyncDate(this.syncId as string, syncStartDate, override);
-            await slackNotificationService.removeFailingConnection(
-                this.nangoConnection,
-                this.syncName,
-                this.syncType,
-                this.activityLogId as number,
-                this.nangoConnection.environment_id,
-                this.provider as string
-            );
+            if (!this.isWebhook) {
+                const override = false;
+                await setLastSyncDate(this.syncId as string, syncStartDate, override);
+                await slackNotificationService.removeFailingConnection(
+                    this.nangoConnection,
+                    this.syncName,
+                    this.syncType,
+                    this.activityLogId as number,
+                    this.nangoConnection.environment_id,
+                    this.provider as string
+                );
+            }
         }
 
-        if (trackDeletes) {
+        if (!this.isWebhook && trackDeletes) {
             await takeSnapshot(this.nangoConnection?.id as number, model);
         }
 
@@ -662,14 +671,16 @@ export default class SyncRun {
             return;
         }
 
-        await slackNotificationService.reportFailure(
-            this.nangoConnection,
-            this.syncName,
-            this.syncType,
-            this.activityLogId as number,
-            this.nangoConnection.environment_id,
-            this.provider as string
-        );
+        if (!this.isWebhook) {
+            await slackNotificationService.reportFailure(
+                this.nangoConnection,
+                this.syncName,
+                this.syncType,
+                this.activityLogId as number,
+                this.nangoConnection.environment_id,
+                this.provider as string
+            );
+        }
 
         if (!this.activityLogId || !this.syncJobId) {
             console.error(content);
@@ -716,5 +727,15 @@ export default class SyncRun {
             },
             `syncId:${this.syncId}`
         );
+    }
+
+    private determineErrorType(): string {
+        if (this.isAction) {
+            return 'action_script_failure';
+        } else if (this.isWebhook) {
+            return 'webhook_script_failure';
+        } else {
+            return 'sync_script_failure';
+        }
     }
 }
